// src/openai/model.types.ts
export type TextModel =
  | 'gpt-5.1'
  | 'gpt-5'
  | 'gpt-5-mini'
  | 'gpt-5-nano'
  | 'gpt-5.1-chat-latest'
  | 'gpt-5-chat-latest'
  | 'gpt-5.1-codex-max'
  | 'gpt-5.1-codex'
  | 'gpt-5-codex'
  | 'gpt-5-pro'
  | 'gpt-4.1'
  | 'gpt-4.1-mini'
  | 'gpt-4.1-nano'
  | 'gpt-4o'
  | 'gpt-4o-2024-05-13'
  | 'gpt-4o-mini'
  | 'gpt-realtime'
  | 'gpt-realtime-mini'
  | 'gpt-4o-realtime-preview'
  | 'gpt-4o-mini-realtime-preview'
  | 'gpt-audio'
  | 'gpt-audio-mini'
  | 'gpt-4o-audio-preview'
  | 'gpt-4o-mini-audio-preview'
  | 'o1'
  | 'o1-pro'
  | 'o3-pro'
  | 'o3'
  | 'o3-deep-research'
  | 'o4-mini'
  | 'o4-mini-deep-research'
  | 'o3-mini'
  | 'o1-mini'
  | 'gpt-5.1-codex-mini'
  | 'codex-mini-latest'
  | 'gpt-5-search-api'
  | 'gpt-4o-mini-search-preview'
  | 'gpt-4o-search-preview'
  | 'computer-use-preview'
  | 'gpt-image-1'
  | 'gpt-image-1-mini';